requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
celery==5.3.6
stripe==7.10.0
qrcode==7.4.2
pillow==10.2.0
requests==2.31.0
httpx==0.26.0
python-dotenv==1.0.0
email-validator==2.1.0
pyotp==2.9.0
phonenumbers==8.13.27
wechatpy>=1.8.18
cryptography==42.0.0
boto3==1.34.0
botocore==1.34.0

# OTP Providers
twilio==9.0.0
sendgrid==6.11.0

# Music API
aiohttp==3.9.1

# OCR for manual payments
pytesseract==0.3.10
